requirements.txt
python-telegram-bot==20.9
matplotlib==3.9.0
pillow==10.4.0
python-dotenv==1.0.1